move tests to mha_dense_tests.cpp
